+++ src/components/feedback/LoadingSpinner.vue
