+++ src/components/navigation/CategorySelector.vue
